introduce growi bridge service
